WIP refactor knockoff module
