+++ hidimstat/__init__.py
@@ -1,9 +1,11 @@
-from .clustered_inference import clustered_inference, hd_inference
-from .desparsified_lasso import desparsified_lasso, desparsified_group_lasso
-from .ensemble_clustered_inference import ensemble_clustered_inference
 from .adaptive_permutation_threshold import ada_svr
+from .clustered_inference import clustered_inference, hd_inference
+from .desparsified_lasso import desparsified_group_lasso, desparsified_lasso
+from .ensemble_clustered_inference import ensemble_clustered_inference
+from .knockoffs import model_x_knockoff
+from .knockoff_aggregation import knockoff_aggregation
 from .multi_sample_split import aggregate_quantiles
-from .noise_std import reid, group_reid
+from .noise_std import group_reid, reid
 from .permutation_test import permutation_test_cv
 from .scenario import multivariate_1D_simulation
 from .standardized_svr import standardized_svr
@@ -11,14 +13,16 @@ from .stat_tools import zscore_from_pval
 from .version import __version__
 
 __all__ = [
+    'ada_svr',
     'aggregate_quantiles',
     'clustered_inference',
     'desparsified_lasso',
     'desparsified_group_lasso',
     'ensemble_clustered_inference',
-    'ada_svr',
     'group_reid',
     'hd_inference',
+    'knockoff_aggregation',
+    'model_x_knockoff',
     'multivariate_1D_simulation',
     'permutation_test_cv',
     'reid',
