--- hidimstat/knockoffs/__init__.py
@@ -1,12 +1,0 @@
-from .gaussian_knockoff import gaussian_knockoff_generation
-from .knockoffs import model_x_knockoff
-from .knockoff_aggregation import knockoff_aggregation
-from .stat_coef_diff import stat_coef_diff
-
-
-__all__ = [
-    'gaussian_knockoff_generation',
-    'knockoff_aggregation',
-    'model_x_knockoff',
-    'stat_coef_diff',
-]
